+++ src/tokens/ERC721/interfaces/IERC721Press.sol
@@ -156,7 +156,7 @@ interface IERC721Press {
     ) external;
 
     /// @notice allows user to mint token(s) from the Press contract
-    function mintWithData(address recipient, uint64 mintQuantity, bytes memory mintData)
+    function mintWithData(address recipient, uint16 mintQuantity, bytes memory mintData)
         external
         payable
         returns (uint256);
